Move stuff from _create_ca to engine (_create_primary_ca)
Things that are engine-specific, should not be in _create_ca which is
used also for qemu ca.

TODO: Need to verify _CA_CERT_CONF _CERT_CONF. I removed them
completely, I think they are already handled now by _update_templates,
which now writes also _CONF, not only _TEMPLATE. This is not really
needed, because pki-create-ca.sh does that too, but should be harmless.

Change-Id: Iaa9a64317b4df812dc77d5551bcca465b639b4e2
Signed-off-by: Yedidyah Bar David <didi@redhat.com>