packaging: pki: Support passing --ca-file to pki-enroll-pkcs12.sh
This is currently not needed, and I am adding it mainly for debugging.

Can be useful in the future, if we want a single command to create
and sign a keypair on the engine machine using the qemu ca. Without
this patch, this requires creating the private key and request
using other means (e.g. calling openssl directly) and then calling
pki-enroll-request.sh.

Change-Id: Icafc99a35545d9514540e3dd0084d3eedc5859ad
Signed-off-by: Yedidyah Bar David <didi@redhat.com>